Added binary format; Added moar tests; fixed some bugs

+++ fast-ftrl-proximal/src/main/java/io/scaledml/core/inputformats/CSVFormat.java
@@ -4,7 +4,7 @@ import com.google.inject.Inject;
 import com.google.inject.name.Named;
 import io.scaledml.core.SparseItem;
 import io.scaledml.ftrl.featuresprocessors.FeaturesProcessor;
-import io.scaledml.ftrl.util.LineBytesBuffer;
+import io.scaledml.core.util.LineBytesBuffer;
 
 /**
  * @author Ilya Smagin ilya-sm@yandex-team.ru on 4/2/15.
@@ -45,7 +45,7 @@ public class CSVFormat implements InputFormat {
             return;
         }
         namespaceBuffer.clear();
-        namespaceBuffer.putInteger(colNum);
+        namespaceBuffer.putShort((short) colNum);
         ColumnsMask.ColumnType columnType = columnsMask.getCategory(colNum);
         switch (columnType) {
             case LABEL:
